change setters for width and height to accept float

+++ src/main/java/com/example/tiledmaptest/Entity.java
@@ -53,7 +53,7 @@ public class Entity {
         return width;
     }
 
-    public void setWidth(final int width) {
+    public void setWidth(final float width) {
         this.width = width;
     }
 
@@ -61,7 +61,7 @@ public class Entity {
         return height;
     }
 
-    public void setHeight(final int height) {
+    public void setHeight(final float height) {
         this.height = height;
     }
 }
